Add sections to homepage and organize static output
