--- cms/src/collections/Homepage.ts
@@ -1,17 +1,0 @@
-import type { GlobalConfig } from 'payload'
-
-export const Homepage: GlobalConfig = {
-  slug: 'homepage',
-  fields: [
-    {
-      name: 'headline',
-      type: 'text',
-      required: true,
-    },
-    {
-      name: 'paragraph',
-      type: 'textarea',
-      required: true,
-    },
-  ],
-}
